requirements.txt
opencv-python
cmake
pyyaml
streamlit==1.22.0
streamlit
opencv-python-headless
numpy
torch
torchvision
scikit-image
pillow
scikit-learn
#Fork of face_recognition using prebuild dlib library instead
face_recognition @ git+https://github.com/thetoby9944/face_recognition

